0.9.109
23/10/2025

+++ src/Z2S_sensor/Z2S_remote_thermometer.h
@@ -207,7 +207,7 @@ public:
       }
     }
     if (connected_thermometers_number > 0) 
-      setValue(connected_thermometers_calculated_temperature / 
+      setValue((double)(connected_thermometers_calculated_temperature) / 
                (connected_thermometers_number * 100));
   }
